refactoring: extracted accessory classes
