+++ index.js
@@ -1,6 +1,11 @@
 var request = require("request");
 var Accessory, Service, Characteristic, UUIDGen;
 
+var RademacherBlindsAccessory = require('./accessories/RademacherBlindsAccessory.js');
+var RademacherLockAccessory = require('./accessories/RademacherLockAccessory.js');
+var RademacherDimmerAccessory = require ('./accessories/RademacherDimmerAccessory.js');
+var RademacherSwitchAccessory = require ('./accessories/RademacherSwitchAccessory.js');
+
 module.exports = function(homebridge) {
     Accessory = homebridge.platformAccessory;
     Service = homebridge.hap.Service;
@@ -123,7 +128,7 @@ RademacherHomePilot.prototype.addDimmerAccessory = function(dimmer) {
         name = dimmer.description;
     var accessory = new Accessory(name, UUIDGen.generate("did"+dimmer.did));
     accessory.addService(Service.Lightbulb, name);
-    this.accessories[accessory.UUID] = new RademacheDimmerAccessory(this.log, accessory, dimmer, this.url, this.inverted);
+    this.accessories[accessory.UUID] = new RademacherDimmerAccessory(this.log, accessory, dimmer, this.url, this.inverted);
     this.api.registerPlatformAccessories("homebridge-rademacher-homepilot", "RademacherHomePilot", [accessory]);
 };
 
@@ -163,418 +168,6 @@ RademacherHomePilot.prototype.removeAccessory = function(accessory) {
         }
         this.api.unregisterPlatformAccessories("homebridge-rademacher-homepilot", "RademacherHomePilot", [accessory]);
     }
-};
-
-function RademacherBlindsAccessory(log, accessory, blind, url, inverted) {
-    var self = this;
-
-    var info = accessory.getService(Service.AccessoryInformation);
-
-    accessory.context.manufacturer = "Rademacher";
-    info.setCharacteristic(Characteristic.Manufacturer, accessory.context.manufacturer.toString());
-
-    accessory.context.model = blind.productName;
-    info.setCharacteristic(Characteristic.Model, accessory.context.model.toString());
-
-    accessory.context.serial = blind.serial;
-    info.setCharacteristic(Characteristic.SerialNumber, accessory.context.serial.toString());
-
-    this.inverted = inverted;
-    this.accessory = accessory;
-    this.blind = blind;
-    this.log = log;
-    this.url = url;
-    this.lastPosition = this.inverted ? reversePercentage(this.blind.position) : this.blind.position;
-    this.currentPositionState = 2;
-    this.currentTargetPosition = 100;
-
-    this.service = accessory.getService(Service.WindowCovering);
-
-    this.service
-        .getCharacteristic(Characteristic.CurrentPosition)
-        .setValue(self.inverted ? reversePercentage(self.blind.position) : self.blind.position)
-        .on('get', this.getCurrentPosition.bind(this));
-
-    this.service
-        .getCharacteristic(Characteristic.TargetPosition)
-        .setValue(self.inverted ? reversePercentage(self.blind.position) : self.blind.position)
-        .on('get', this.getTargetPosition.bind(this))
-        .on('set', this.setTargetPosition.bind(this));
-
-    this.service.getCharacteristic(Characteristic.PositionState)
-        .setValue(this.currentPositionState)
-        .on('get', this.getPositionState.bind(this));
-
-    accessory.updateReachability(true);
-}
-
-RademacherBlindsAccessory.prototype.setTargetPosition = function(value, callback) {
-    this.log("%s - Setting target position: %s", this.accessory.displayName, value);
-
-    var self = this;
-    this.currentTargetPosition = value;
-    var moveUp = (this.currentTargetPosition >= this.lastPosition);
-    this.service.setCharacteristic(Characteristic.PositionState, (moveUp ? 1 : 0));
-    var target = self.inverted ? reversePercentage(value) : value;
-
-    var params = "cid=9&did="+this.blind.did+"&command=1&goto="+ target;
-    request.post({
-        headers: {'content-type' : 'application/x-www-form-urlencoded'},
-        url: this.url + "/deviceajax.do",
-        body: params
-    }, function(e,r,b){
-        if(e) return callback(new Error("Request failed."), false);
-        if(r.statusCode == 200)
-        {
-            self.service.setCharacteristic(Characteristic.CurrentPosition, self.currentTargetPosition);
-            self.service.setCharacteristic(Characteristic.PositionState, 2);
-            self.lastPosition = self.currentTargetPosition;
-            callback(null, self.currentTargetPosition);
-        }
-    });
-};
-
-RademacherBlindsAccessory.prototype.getTargetPosition = function(callback) {
-    this.log("%s - Getting target position", this.accessory.displayName);
-
-    var self = this;
-    var did = this.blind.did;
-
-    request.get({
-        timeout: 2500,
-        strictSSL: false,
-        url: this.url + "/deviceajax.do?devices=1"
-    }, function(e,r,b) {
-        if(e) return callback(new Error("Request failed."), false);
-        var body = JSON.parse(b);
-        body.devices.forEach(function(data) {
-            if(data.did == did)
-            {
-                var pos = self.inverted ? reversePercentage(data.position) : data.position;
-                self.currentTargetPosition = pos;
-                callback(null, pos);
-            }
-        });
-    });
-};
-
-RademacherBlindsAccessory.prototype.getCurrentPosition = function(callback) {
-    this.log("%s - Getting current position", this.accessory.displayName);
-
-    var self = this;
-    var did = this.blind.did;
-
-    request.get({
-        timeout: 2500,
-        strictSSL: false,
-        url: this.url + "/deviceajax.do?devices=1"
-    }, function(e,r,b) {
-        if(e) return callback(new Error("Request failed: "+e), false);
-        var body = JSON.parse(b);
-        body.devices.forEach(function(data) {
-            if(data.did == did)
-            {
-                var pos = self.inverted ? reversePercentage(data.position) : data.position;
-                callback(null, pos);
-            }
-        });
-    });
-};
-
-RademacherBlindsAccessory.prototype.getPositionState = function(callback) {
-    callback(null, this.currentPositionState);
-};
-
-function RademacherSwitchAccessory(log, accessory, sw, url) {
-    var self = this;
-
-    var info = accessory.getService(Service.AccessoryInformation);
-
-    accessory.context.manufacturer = "Rademacher";
-    info.setCharacteristic(Characteristic.Manufacturer, accessory.context.manufacturer.toString());
-
-    accessory.context.model = sw.productName;
-    info.setCharacteristic(Characteristic.Model, accessory.context.model.toString());
-
-    accessory.context.serial = sw.serial;
-    info.setCharacteristic(Characteristic.SerialNumber, accessory.context.serial.toString());
-
-    this.accessory = accessory;
-    this.sw = sw;
-    this.log = log;
-    this.url = url;
-    this.lastState = this.sw.position==100?true:false;
-    this.currentState = this.sw.position==100?true:false;
-
-    this.service = accessory.getService(Service.Switch);
-
-    this.service
-        .getCharacteristic(Characteristic.On)
-        .setValue(self.sw.position==100?true:false)
-        .on('set', this.setCurrentState.bind(this))
-        .on('get', this.getCurrentState.bind(this));
-
-    accessory.updateReachability(true);
-}
-
-RademacherSwitchAccessory.prototype.getCurrentState = function(callback) {
-    this.log("%s - Getting current state for %s", this.accessory.displayName, this.accessory.UUID);
-
-    var self = this;
-    var serial = this.sw.serial;
-    var name = this.sw.name;
-
-    request.get({
-        timeout: 1500,
-        strictSSL: false,
-        url: this.url + "/deviceajax.do?devices=1"
-    }, function(e,r,b) {
-        if(e) return callback(new Error("Request failed: "+e), false);
-        var body = JSON.parse(b);
-        body.devices.forEach(function(data) {
-            if(data.serial == serial && data.name == name)
-            {
-                var pos = data.position;
-                callback(null, (pos==100?true:false));
-            }
-        });
-    });
-};
-
-RademacherSwitchAccessory.prototype.setCurrentState = function(value, callback) {
-    this.log("%s - Setting switch: %s", this.accessory.displayName, value);
-
-    var self = this;
-    this.currentState = value;
-    var changed = (this.currentState != this.lastState);
-    this.log("%s - switch changed=%s", this.accessory.displayName, changed);
-    if (changed)
-    {
-       var params = "cid="+(this.currentState?"10":"11")+"&did="+this.sw.did+"&command=1";
-       request.post({
-           headers: {'content-type' : 'application/x-www-form-urlencoded'},
-           url: this.url + "/deviceajax.do",
-           body: params
-       }, function(e,r,b){
-            if(e) return callback(new Error("Request failed."), self.currentState);
-            if(r.statusCode == 200)
-            {
-                self.lastState = self.currentState;
-                return callback(null, self.currentState);
-            }
-            else
-            {
-                return callback(new Error("Request failed with status "+r.statusCode), self.currentState);
-            }
-        });
-    }
-    else
-    {
-        return callback(null,this.currentState);
-    }
-};
-
-function RademacherLockAccessory(log, accessory, sw, url) {
-  this.log = log;
-  this.sw = sw;
-  this.url = url;
-  this.accessory = accessory;
-  this.lockservice = accessory.getService(Service.LockMechanism);
-  
-  this.lockservice
-    .getCharacteristic(Characteristic.LockCurrentState)
-    .on('get', this.getState.bind(this));
-  
-  this.lockservice
-    .getCharacteristic(Characteristic.LockTargetState)
-    .on('get', this.getState.bind(this))
-    .on('set', this.setState.bind(this));
-	
-}
-
-RademacherLockAccessory.prototype.getState = function(callback) {
-    this.log("Get lock state of %s", this.accessory.displayName)
-    callback(null, true);
-}
-
-RademacherLockAccessory.prototype.setState = function(state, callback) {
-    var self=this;
-    var lockState = (state == Characteristic.LockTargetState.SECURED) ? "lock" : "unlock";
-    this.log("Set lock state of %s to %s", this.accessory.displayName,lockState);
-
-    var params = "cid=10&did="+this.sw.did+"&command=1";
-    request.post({
-        headers: {'content-type' : 'application/x-www-form-urlencoded'},
-        url: this.url + "/deviceajax.do",
-        body: params
-        }, function(e,r,b){
-            if(e) return callback(new Error("Request failed."), false);
-            if(r.statusCode == 200)
-            {
-                // alway unlock
-                self.lockservice.setCharacteristic(Characteristic.LockCurrentState, Characteristic.LockCurrentState.UNSECURED);
-                self.lockservice.setCharacteristic(Characteristic.LockCurrentState, Characteristic.LockCurrentState.SECURED)
-                callback(null);
-            }
-    });
-}
-
-
-RademacherLockAccessory.prototype.getServices = function() {
-  return [this.lockservice];
-}
-
-function RademacherDimmerAccessory(log, accessory, dimmer, url, inverted) {
-    var self = this;
-
-    var info = accessory.getService(Service.AccessoryInformation);
-
-    accessory.context.manufacturer = "Rademacher";
-    info.setCharacteristic(Characteristic.Manufacturer, accessory.context.manufacturer.toString());
-
-    accessory.context.model = dimmer.productName;
-    info.setCharacteristic(Characteristic.Model, accessory.context.model.toString());
-
-    accessory.context.serial = dimmer.serial;
-    info.setCharacteristic(Characteristic.SerialNumber, accessory.context.serial.toString());
-
-    this.inverted = inverted;
-    this.accessory = accessory;
-    this.dimmer = dimmer;
-    this.log = log;
-    this.url = url;
-    this.lastBrightness = this.inverted ? reversePercentage(this.dimmer.brightness) : this.dimmer.brightness;
-    this.currentBrightness = 2;
-    this.currentStatus = 100;
-
-    this.service = accessory.getService(Service.Lightbulb);
-
-    this.service.getCharacteristic(Characteristic.On)
-        .on('get', this.getStatus.bind(this))
-        .on('set', this.setStatus.bind(this));
-
-    this.service.getCharacteristic(Characteristic.Brightness)
-        .on('get', this.getBrightness.bind(this))
-        .on('set', this.setBrightness.bind(this));
-
-    accessory.updateReachability(true);
-}
-
-RademacherDimmerAccessory.prototype.isInt = function (value) {
-    return /^-?[0-9]+$/.test(value);
-};
-
-RademacherDimmerAccessory.prototype.getStatus = function (callback) {
-    this.log("%s - Getting current state for %s", this.accessory.displayName, this.accessory.UUID);
-
-    var self = this;
-    var serial = this.dimmer.serial;
-    var name = this.dimmer.name;
-
-    request.get({
-        timeout: 1500,
-        strictSSL: false,
-        url: this.url + "/deviceajax.do?devices=1"
-    }, function(e,r,b) {
-        if(e) return callback(new Error("Request failed: "+e), false);
-        var body = JSON.parse(b);
-        body.devices.forEach(function(data) {
-            if(data.serial == serial && data.name == name)
-            {
-                var pos = data.position;
-                callback(null, (pos>0?true:false));
-            }
-        });
-    });
-};
-
-RademacherDimmerAccessory.prototype.setStatus = function (status, callback, context) {
-    if (context !== 'fromSetValue') {
-        this.on = status;
-        this.log("%s - Setting switch: %s", this.accessory.displayName, status);
-
-        var self = this;
-        this.currentState = status;
-        var changed = (this.currentState != this.lastState);
-        this.log("%s - switch changed=%s", this.accessory.displayName, changed);
-        if (changed)
-        {
-           var params = "cid="+(this.currentState?"10":"11")+"&did="+this.sw.did+"&command=1";
-           request.post({
-               headers: {'content-type' : 'application/x-www-form-urlencoded'},
-               url: this.url + "/deviceajax.do",
-               body: params
-           }, function(e,r,b){
-                if(e) return callback(new Error("Request failed."), self.currentState);
-                if(r.statusCode == 200)
-                {
-                    self.lastState = self.currentState;
-                    return callback(null, self.currentState);
-                }
-                else
-                {
-                    return callback(new Error("Request failed with status "+r.statusCode), self.currentState);
-                }
-            });
-        }
-        else
-        {
-            return callback(null,this.currentState);
-        }
-    }
-};
-
-RademacherDimmerAccessory.prototype.getBrightness = function (callback) {
-    this.log("%s - Getting current brightness", this.accessory.displayName);
-
-    var self = this;
-    var did = this.dimmer.did;
-
-    request.get({
-        timeout: 2500,
-        strictSSL: false,
-        url: this.url + "/deviceajax.do?devices=1"
-    }, function(e,r,b) {
-        if(e) return callback(new Error("Request failed: "+e), false);
-        var body = JSON.parse(b);
-        body.devices.forEach(function(data) {
-            if(data.did == did)
-            {
-                var pos = self.inverted ? reversePercentage(data.position) : data.position;
-                callback(null, pos);
-            }
-        });
-    });
-};
-
-RademacherDimmerAccessory.prototype.setBrightness = function (brightness, callback, context) {
-    if (context !== 'fromSetValue') {
-        this.log("%s - Setting target brightness: %s", this.accessory.displayName, brightness);
-        var self = this;
-        this.currentBrightness = brightness;
-        var moveUp = (this.currentBrightness >= this.lastBrightness);
-        this.service.setCharacteristic(Characteristic.Brightness, (moveUp ? 1 : 0));
-        var target = self.inverted ? reversePercentage(brightness) : brightness;
-    
-        var params = "cid=9&did="+this.blind.did+"&command=1&goto="+ target;
-        request.post({
-            headers: {'content-type' : 'application/x-www-form-urlencoded'},
-            url: this.url + "/deviceajax.do",
-            body: params
-        }, function(e,r,b){
-            if(e) return callback(new Error("Request failed."), false);
-            if(r.statusCode == 200)
-            {
-                self.service.setCharacteristic(Characteristic.Brightness, self.currentBrightness);
-                self.lastBrightness = self.currentBrightness;
-                callback(null, self.currentBrightness);
-            }
-        });
-    }
-};
-
-RademacherDimmerAccessory.prototype.getServices = function () {
-    return [this.service];
 };
 
 function reversePercentage(p) {
